Switch to new Brooklyn package names.
